Move `Quoting` resource to dbt/artifacts
